// types/bang.ts
export interface Bang {
  c: string // category
  d: string // domain
  r: number // rank
  s: string // site name
  sc: string // subcategory
  t: string // trigger
  u: string // url
}

export interface BangData {
  trigger: string
  name: string
  url: string
  category: string
  subcategory: string
  domain: string
}